+++ waspc/cli/src/Wasp/Cli/Command/BuildStart/Config.hs
@@ -11,50 +11,44 @@ import Control.Monad.Except (MonadError (throwError))
 import Control.Monad.IO.Class (MonadIO (liftIO))
 import Data.Char (toLower)
 import Data.List (intercalate)
+import Network.Socket (PortNumber)
 import StrongPath ((</>))
 import qualified StrongPath as SP
 import Wasp.AppSpec (AppSpec)
 import qualified Wasp.AppSpec.Valid as ASV
 import Wasp.Cli.Command (Command, CommandError (CommandError))
 import Wasp.Cli.Command.BuildStart.ArgumentsParser (BuildStartArgs (..), buildStartArgsParser)
-import Wasp.Cli.Util.AppSides (AppSides (..))
-import qualified Wasp.Cli.Util.AppSides as AppSides
+import Wasp.Cli.Util.Apps (getDevUrlMakers, getWaspEnvVars)
 import Wasp.Cli.Util.Parser (getParserHelpMessage)
 import Wasp.Cli.Util.PathArgument (FilePathArgument)
 import qualified Wasp.Cli.Util.PathArgument as PathArgument
+import qualified Wasp.Cli.Util.PortArgument as PortArgument
 import Wasp.Env (EnvVar, nubEnvVars, overrideEnvVars, parseDotEnvFile)
 import Wasp.Generator.Common (GeneratedAppDir)
-import qualified Wasp.Generator.ServerGenerator.Common as Server
-import qualified Wasp.Generator.WebAppGenerator.Common as WebApp
+import Wasp.Project.Apps (Apps (..))
 import Wasp.Project.Common (WaspProjectDir, generatedAppDirInWaspProjectDir, makeAppUniqueId)
 import Wasp.Util.Terminal (styleCode)
 
 data BuildStartConfig = BuildStartConfig
   { appUniqueId :: String,
-    ports :: AppSides Int,
-    urls :: AppSides String,
-    envVars :: AppSides [EnvVar],
+    ports :: Apps PortNumber,
+    urls :: Apps String,
+    envVars :: Apps [EnvVar],
     buildDir :: SP.Path' SP.Abs (SP.Dir GeneratedAppDir),
     projectDir :: SP.Path' SP.Abs (SP.Dir WaspProjectDir)
   }
 
 makeBuildStartConfig :: AppSpec -> BuildStartArgs -> SP.Path' SP.Abs (SP.Dir WaspProjectDir) -> Command BuildStartConfig
 makeBuildStartConfig appSpec args projectDir' = do
-  userEnvVars <- traverse (liftIO . combineEnvVarsWithEnvFiles) args.envInputs
+  ports <- PortArgument.resolveAppPorts args.ports
+
+  userEnvVars <- liftIO $ traverse combineEnvVarsWithEnvFiles args.envInputs
   when (all null userEnvVars) $ throwError noEnvVarsSpecifiedMsg
 
-  let waspEnvVars =
-        AppSides
-          { client =
-              [ (WebApp.serverUrlEnvVarName, urls.server)
-              ],
-            server =
-              [ (Server.clientUrlEnvVarName, urls.client),
-                (Server.serverUrlEnvVarName, urls.server)
-              ]
-          }
+  let urls = getDevUrlMakers appSpec <*> ports
+      waspEnvVars = getWaspEnvVars appSpec ports
 
-  envVars <- sequenceA $ liftA2 overrideEnvVarsCommand waspEnvVars userEnvVars
+  envVars <- sequence $ liftA2 overrideEnvVarsCommand waspEnvVars userEnvVars
 
   return $
     BuildStartConfig
@@ -70,15 +64,6 @@ makeBuildStartConfig appSpec args projectDir' = do
     (appName, _) = ASV.getApp appSpec
 
     buildDir' = projectDir' </> generatedAppDirInWaspProjectDir
-
-    -- NOTE(carlos): For now, creating these URLs and ports below uses the default
-    -- values we've hardcoded in the generator. In the future, we might want to make
-    -- these configurable via the Wasp app spec or command line arguments.
-
-    -- This assumes that `getDefaultDevClientUrl` uses `defaultClientPort` internally.
-    -- If that changes, we also need to change this.
-    ports = AppSides.defaultPorts
-    urls = AppSides.defaultDevUrls appSpec
 
     noEnvVarsSpecifiedMsg =
       CommandError