+++ waspc/cli/src/Wasp/Cli/Util/Apps.hs
@@ -1,0 +1,26 @@
+module Wasp.Cli.Util.Apps where
+
+import Network.Socket (PortNumber)
+import Wasp.AppSpec (AppSpec)
+import Wasp.Env (EnvVar)
+import qualified Wasp.Generator.ServerGenerator.Common as Server
+import qualified Wasp.Generator.WebAppGenerator.Common as WebApp
+import Wasp.Project.Apps (Apps (..))
+
+getDevUrlMakers :: AppSpec -> Apps (PortNumber -> String)
+getDevUrlMakers spec =
+  Apps
+    { client = WebApp.getDevClientUrl spec,
+      server = Server.makeDevServerUrl
+    }
+
+getWaspEnvVars :: AppSpec -> Apps PortNumber -> Apps [EnvVar]
+getWaspEnvVars spec ports =
+  Apps
+    { client = WebApp.getDevClientEnvVars,
+      server = Server.getDevServerEnvVars
+    }
+    <*> pure locations
+  where
+    locations = liftA2 (,) ports urls
+    urls = getDevUrlMakers spec <*> ports